♻️ Refactor shortcut constants and improve settings UI
